上传插件

+++ hprofanalyzer/src/main/java/com/source/hprofanalyzer/util/SizeUnit.kt
@@ -23,9 +23,10 @@ sealed class SizeUnit {
   abstract fun toKB(value: Long): Float
   abstract fun toMB(value: Long): Float
 
-  abstract fun toByte(value: Float): Float
+  abstract fun toByte(value: Int): Float
   abstract fun toKB(value: Int): Float
   abstract fun toMB(value: Int): Float
+
 
   object BYTE : SizeUnit() {
     override fun toByte(value: Long) = value.toFloat()
@@ -34,7 +35,7 @@ sealed class SizeUnit {
 
     override fun toMB(value: Long) = value / 1024.0f / 1024.0f
 
-    override fun toByte(value: Float) = value.toFloat()
+    override fun toByte(value: Int) = value.toFloat()
 
     override fun toKB(value: Int) = value / 1024.0f
 
@@ -48,7 +49,7 @@ sealed class SizeUnit {
 
     override fun toMB(value: Long) = value / 1024.0f
 
-    override fun toByte(value: Float) = value * 1024.0f
+    override fun toByte(value: Int) = value * 1024.0f
 
     override fun toKB(value: Int) = value.toFloat()
 
@@ -62,7 +63,7 @@ sealed class SizeUnit {
 
     override fun toMB(value: Long) = value.toFloat()
 
-    override fun toByte(value: Float) = value * 1024.0f * 1024.0f
+    override fun toByte(value: Int) = value * 1024.0f * 1024.0f
 
     override fun toKB(value: Int) = value * 1024.0f
